requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
fastapi==0.115.12
filelock==3.18.0
h11==0.14.0
idna==3.10
joblib==1.4.2
mpmath==1.3.0
mysql-connector-python==9.3.0
numpy==2.2.4
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.1
virtualenv==20.30.0
wheel==0.45.1
